refactor: move kv to separate file
